+++ src/battle-main.h
@@ -1,3 +1,5 @@
-unsigned long int battleCount;
+#include <stdint.h>
+
+uint16_t battleCount;
 
 void execBattle(void);
